Modifications in API's

+++ src/controllers/accountManagement.js
@@ -1,9 +1,9 @@
 const axios = require("axios");
 const { v4: uuidv4 } = require("uuid");
 const { responseCode, rs } = require("../util");
+const User = require("./../models/userAuth");
+const {common } = require("../util/helper");
 
-
-let userToken = process.env.USER_AUTH_TOKEN;
 let baseUrl = process.env.SFOX_BASE_URL;
 
 
@@ -25,12 +25,25 @@ exports.linkBank = async (req, res) => {
     };
 
   try {
+    const {user_id} = req.params;
+    const count = await User.scan().exec()
+    var getUser = count.filter((item) => item.user_id == user_id);
+
+    if(getUser.length == 0 || getUser[0].userToken == ""){
+      common.eventBridge(
+        "USER NOT FOUND",
+        responseCode.badRequest
+      );
+      return res
+        .status(responseCode.badRequest)
+        .json(rs.incorrectDetails("USER NOT FOUND", {}));
+    }
     const apiPath = `${baseUrl}/v1/user/bank`;
     let response = await axios({
       method: "post",
       url: apiPath,
       headers: {
-        Authorization: "Bearer " + userToken,
+        Authorization: "Bearer " + getUser[0].userToken,
       },
       data: data,
     });
@@ -41,12 +54,25 @@ exports.linkBank = async (req, res) => {
     console.log("error", err);
     return res
       .status(err.response?.status)
-      .json({ error: err.response.data.error });
+      .json({ error: err.response?.data?.error });
   }
 };
 
 exports.verifyBank = async (req, res) => {
   try {
+    const {user_id} = req.params;
+    const count = await User.scan().exec()
+    var getUser = count.filter((item) => item.user_id == user_id);
+
+    if(getUser.length == 0 || getUser[0].userToken == ""){
+      common.eventBridge(
+        "USER NOT FOUND",
+        responseCode.badRequest
+      );
+      return res
+        .status(responseCode.badRequest)
+        .json(rs.incorrectDetails("USER NOT FOUND", {}));
+    }
     let data = {
       amount1: req.body.verifyAmount1,
       amount2: req.body.verifyAmount2,
@@ -56,61 +82,100 @@ exports.verifyBank = async (req, res) => {
       method: "post",
       url: apiPath,
       headers: {
-        Authorization: "Bearer " + userToken,
+        Authorization: "Bearer " + getUser[0].userToken,
       },
       data: data,
     });
     return res.status(response.status).json({ message: response.data.data });
   } catch (err) {
-    return res.status(err.response?.status).send(err.response.data);
+    return res.status(err.response?.status).send(err.response?.data);
   }
 };
 
 exports.getBank = async (req, res) => {
   try {
+    const {user_id} = req.params;
+    const count = await User.scan().exec()
+    var getUser = count.filter((item) => item.user_id == user_id);
+
+    if(getUser.length == 0 || getUser[0].userToken == ""){
+      common.eventBridge(
+        "USER NOT FOUND",
+        responseCode.badRequest
+      );
+      return res
+        .status(responseCode.badRequest)
+        .json(rs.incorrectDetails("USER NOT FOUND", {}));
+    }
     let apiPath = `${baseUrl}/v1/user/bank`;
     let response = await axios({
       method: "get",
       url: apiPath,
       headers: {
-        Authorization: "Bearer " + userToken,
+        Authorization: "Bearer " + getUser[0].userToken,
       },
     });
     return res.status(response.status).json({ message: response.data });
   } catch (err) {
-    return res.status(err.response?.status).send(err.response.data);
+    return res.status(err.response?.status).send(err.response?.data);
   }
 };
 
 exports.deleteBank = async (req, res) => {
   try {
+    const {user_id} = req.params;
+    const count = await User.scan().exec()
+    var getUser = count.filter((item) => item.user_id == user_id);
+
+    if(getUser.length == 0 || getUser[0].userToken == ""){
+      common.eventBridge(
+        "USER NOT FOUND",
+        responseCode.badRequest
+      );
+      return res
+        .status(responseCode.badRequest)
+        .json(rs.incorrectDetails("USER NOT FOUND", {}));
+    }
     let apiPath = `${baseUrl}/v1/user/bank`;
     let response = await axios({
       method: "delete",
       url: apiPath,
       headers: {
-        Authorization: "Bearer " + userToken,
+        Authorization: "Bearer " + getUser[0].userToken,
       },
     });
     return res.status(response.status).json({ message: "BANK ACCOUNT DELTED SUCCESSFULLY"});
   } catch (err) {
-    return res.status(err.response?.status).send(err.response.data);
+    return res.status(err.response?.status).send(err.response?.data);
   }
 };
 
 exports.wireInstructions = async (req, res) => {
   try {
+    const {user_id} = req.params;
+    const count = await User.scan().exec()
+    var getUser = count.filter((item) => item.user_id == user_id);
+
+    if(getUser.length == 0 || getUser[0].userToken == ""){
+      common.eventBridge(
+        "USER NOT FOUND",
+        responseCode.badRequest
+      );
+      return res
+        .status(responseCode.badRequest)
+        .json(rs.incorrectDetails("USER NOT FOUND", {}));
+    }
     let apiPath = `${baseUrl}/v1/user/wire-instructions`;
     let response = await axios({
       method: "get",
       url: apiPath,
       headers: {
-        Authorization: "Bearer " + userToken,
+        Authorization: "Bearer " + getUser[0].userToken,
       },
     });
     return res.status(response.status).json({ message: response.data.data });
   } catch (err) {
-    return res.status(err.response?.status).send(err.response.data);
+    return res.status(err.response?.status).send(err.response?.data);
   }
 };
